restore /Canvas/test-suites/JS to Jenkinsfile
fixes Jest tests in the build

also:
  - runs RCE tests in parallel
  - splits IconMakerTray.test.js into two files (large; was fickle)
    - uses .js extension for this file due to babel config

Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/367692
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Aaron Shafovaloff <ashafovaloff@instructure.com>
Product-Review: Aaron Shafovaloff <ashafovaloff@instructure.com>
Build-Review: James Butters <jbutters@instructure.com>
