+++ packages/canvas-rce/package.json
@@ -10,10 +10,10 @@
     "integration-test": "nightwatch --env integration",
     "lint": "eslint --quiet",
     "lint:fix": "eslint --quiet --fix",
-    "test": "yarn test:serial",
+    "test": "yarn jest",
     "test:parallel": "yarn test:jest",
     "test:serial": "yarn test:jest",
-    "test:jest": "jest --color --runInBand",
+    "test:jest": "jest --color",
     "test:cafe": "yarn build:cafe && yarn test:cafe:only",
     "test:cafe:ui": "yarn build:cafe && yarn test:cafe:only:ui",
     "test:cafe:only": "testcafe ${TESTCAFE_PROVIDER:-chrome:headless} testcafe/**/*.test.js",
